+++ app/src/main/java/edu/up/cs/androidcatan/catan/CatanHumanPlayer.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan.catan.players;
+package edu.up.cs.androidcatan.catan;
 
 
 import android.view.View;
