+++ app/src/main/java/edu/up/cs/androidcatan/catan/gamestate/Dice.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan.catan;
+package edu.up.cs.androidcatan.catan.gamestate;
 
 /**
  * @author Alex Weininger
