+++ app/src/main/java/edu/up/cs/androidcatan/catan/gamestate/buildings/Settlement.java
@@ -1,4 +1,4 @@
-package edu.up.cs.androidcatan.catan.board.buildings;
+package edu.up.cs.androidcatan.catan.gamestate.buildings;
 
 import java.util.Arrays;
 
@@ -10,26 +10,26 @@ import java.util.Arrays;
  * @version October 31, 2018
  * https://github.com/alexweininger/android-catan
  **/
-public class City extends Building {
+public class Settlement extends Building {
 
-    public static final int[] resourceCost = {0, 2, 0, 3, 0}; // Brick, Grain, Lumber, Ore, Wool
+    public final static int[] resourceCost = {1, 1, 1, 0, 1}; // Brick, Grain, Lumber, Ore, Wool
 
     /**
-     * @param ownerId id of who owns the building
+     * @param ownerId - player id of who owns the settlement
      */
-    public City(int intersectionId, int ownerId) {
+    public Settlement(int ownerId) {
         super(ownerId);
+    } // end constructor
+
+    public int getVictoryPoints() {
+        return 1;
     }
 
-    @Override
-    public int getVictoryPoints() {
-        return 2;
-    }
 
     @Override
     public String toString() {
-        return "City{" +
+        return "Settlement{" +
                 "resourceCost=" + Arrays.toString(resourceCost) +
                 '}';
     }
-}
+} // end Class
